+++ evaluator.go
@@ -9,9 +9,13 @@ import (
 
 func evaluate(path string, info os.FileInfo, n *tnode) bool {
 	if n.ntype == T_STARTSWITH {
-		return startsWith(path, left(n), info, right(n).sval)
+		return startsWith(path, left(n), info, right(n).sval, true)
+	} else if n.ntype == T_ICSTARTSWITH {
+		return startsWith(path, left(n), info, right(n).sval, false)
 	} else if n.ntype == T_ENDSWITH {
-		return endsWith(path, left(n), info, right(n).sval)
+		return endsWith(path, left(n), info, right(n).sval, true)
+	} else if n.ntype == T_ICENDSWITH {
+		return endsWith(path, left(n), info, right(n).sval, false)
 	} else if n.ntype == T_GT {
 		if resolveAsInt(left(n), info) > resolveAsInt(right(n), info) {
 			return true
@@ -42,6 +46,16 @@ func evaluate(path string, info os.FileInfo, n *tnode) bool {
 				return true
 			}
 		}
+	} else if n.ntype == T_ICEQ {
+		if left(n).ntype == T_NAME {
+			if strings.ToLower(resolveAsString(path, left(n), info)) == strings.ToLower(right(n).sval) {
+				return true
+			} else if left(n).ntype == T_PATH {
+				if strings.ToLower(resolveAsString(path, left(n), info)) == strings.ToLower(right(n).sval) {
+					return true
+				}
+			}
+		}
 	} else if n.ntype == T_NEQ {
 		if left(n).ntype == T_NAME {
 			if resolveAsString(path, left(n), info) != right(n).sval {
@@ -53,6 +67,16 @@ func evaluate(path string, info os.FileInfo, n *tnode) bool {
 			}
 		} else {
 			if resolveAsInt(left(n), info) != resolveAsInt(right(n), info) {
+				return true
+			}
+		}
+	} else if n.ntype == T_ICNEQ {
+		if left(n).ntype == T_NAME {
+			if strings.ToLower(resolveAsString(path, left(n), info)) != strings.ToLower(right(n).sval) {
+				return true
+			}
+		} else if left(n).ntype == T_PATH {
+			if strings.ToLower(resolveAsString(path, left(n), info)) != strings.ToLower(right(n).sval) {
 				return true
 			}
 		}
@@ -71,7 +95,9 @@ func evaluate(path string, info os.FileInfo, n *tnode) bool {
 	} else if n.ntype == T_ISDIR {
 		return info.IsDir()
 	} else if n.ntype == T_CONTAINS {
-		return contains(left(n).ntype, right(n).sval, path, info)
+		return contains(left(n).ntype, right(n).sval, path, info, true)
+	} else if n.ntype == T_ICCONTAINS {
+		return contains(left(n).ntype, right(n).sval, path, info, false)
 	}
 	return false
 }
@@ -112,29 +138,58 @@ func resolveAsInt(n *tnode, info os.FileInfo) int {
 	return 0
 }
 
-func contains(ntype int, search string, path string, info os.FileInfo) bool {
+func contains(ntype int, search string, path string, info os.FileInfo, caseSensitive bool) bool {
+	if !caseSensitive {
+		search = strings.ToLower(search)
+	}
+
 	if ntype == T_NAME {
-		return strings.Contains(info.Name(), search)
+		var name string
+		if caseSensitive {
+			name = info.Name()
+		} else {
+			name = strings.ToLower(info.Name())
+		}
+		return strings.Contains(name, search)
 	} else if ntype == T_PATH {
+		if !caseSensitive {
+			path = strings.ToLower(path)
+		}
 		return strings.Contains(path, search)
 	} else if ntype == T_CONTENT {
-		return !info.IsDir() && fileContainsString(path, search)
+		return !info.IsDir() && fileContainsString(path, search, caseSensitive)
 	}
 	return false
 }
 
-func startsWith(path string, n *tnode, info os.FileInfo, search string) bool {
+func startsWith(path string, n *tnode, info os.FileInfo, search string, caseSensitive bool) bool {
+	if !caseSensitive {
+		search = strings.ToLower(search)
+	}
+
 	if n.ntype == T_CONTENT {
-		return fileStartsWithString(path, search)
+		return fileStartsWithString(path, search, caseSensitive)
 	} else {
-		return strings.HasPrefix(resolveAsString(path, n, info), search)
+		operandValue := resolveAsString(path, n, info)
+		if !caseSensitive {
+			operandValue = strings.ToLower(operandValue)
+		}
+		return strings.HasPrefix(operandValue, search)
 	}
 }
 
-func endsWith(path string, n *tnode, info os.FileInfo, search string) bool {
+func endsWith(path string, n *tnode, info os.FileInfo, search string, caseSensitive bool) bool {
+	if !caseSensitive {
+		search = strings.ToLower(search)
+	}
+
 	if n.ntype == T_CONTENT {
-		return fileEndsWithString(path, info, search)
+		return fileEndsWithString(path, info, search, caseSensitive)
 	} else {
-		return strings.HasSuffix(resolveAsString(path, n, info), search)
+		operandValue := resolveAsString(path, n, info)
+		if !caseSensitive {
+			operandValue = strings.ToLower(operandValue)
+		}
+		return strings.HasSuffix(operandValue, search)
 	}
 }